Feat: Add full CI/CD automation and quality improvements
This commit introduces a comprehensive set of improvements to the repository's CI/CD pipeline, test suite, and documentation, preparing it for the v0.2.0 release.

- **CI/CD and Automation:**
  - Adds a new GitHub Actions workflow (`publish.yml`) to fully automate building, publishing to PyPI, and creating GitHub Releases on new version tags.
  - Adds a Dependabot configuration (`.github/dependabot.yml`) for automated dependency updates.
  - Integrates Codecov for test coverage reporting.
  - Enhances the `ci.yml` workflow to include steps for building the package and documentation.

- **Fixes:**
  - Resolves CI build failures by removing Python 3.13 from the test matrix and pinning the `mtflib` dependency to `v1.5.1`.
  - Corrects a path issue in `docs/build_docs.sh`.
  - Fixes a `TypeError` in `em_app/plotting.py` when generating heatmaps.

- **Testing:**
  - Adds a new test file (`tests/test_plotting.py`) to provide coverage for the plotting module.
  - Optimizes plotting tests to ensure they run quickly and reliably in the CI environment.

- **Documentation:**
  - Adds a `CHANGELOG.md` to track notable changes for each release.
  - Adds and updates `RELEASE.md` to reflect the new, fully automated release process.

+++ src/em_app/plotting.py
@@ -299,17 +299,17 @@ def plot_2d_field(
             field_data = np.array([
                 b.x.extract_coefficient(tuple([0] * b.x.dimension)).item()
                 for b in vector_field._vectors_mtf
-            ], dtype=float)
+            ])
         elif field_component == "y":
             field_data = np.array([
                 b.y.extract_coefficient(tuple([0] * b.y.dimension)).item()
                 for b in vector_field._vectors_mtf
-            ], dtype=float)
+            ])
         else:  # z
             field_data = np.array([
                 b.z.extract_coefficient(tuple([0] * b.z.dimension)).item()
                 for b in vector_field._vectors_mtf
-            ], dtype=float)
+            ])
         field_data = np.real(field_data)
         c = ax.pcolormesh(A, B, field_data.reshape(A.shape), **kwargs)
         plt.colorbar(c, ax=ax)
